backend/requirements.txt
boto3>=1.28.0
pandas>=2.0.0
numpy>=1.26.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
chardet>=5.2.0
openpyxl>=3.1.0
gunicorn>=21.2.0